+++ lib/hexabase.dart
@@ -2,3 +2,4 @@ library hexabase;
 
 export 'src/hexabase.dart';
 export 'src/env.dart';
+export 'src/application.dart';