+++ test/hexabase_test.dart
@@ -13,34 +13,23 @@ void main() {
         : await File('./$path').readAsString();
     var keys = json.decode(str);
     var client = Hexabase();
-    await client.auth.signIn(keys['email'], keys['password']);
+    await client.login(keys['email'], keys['password']);
   });
 
   test('Not login error', () async {
     var client = Hexabase();
     try {
-      await client.workspace.all();
+      await client.workspaces();
       expect(true, false);
     } catch (e) {
       expect(e.toString(), contains('Not authenticated'));
     }
   });
-  test('No initializing error', () async {
-    try {
-      var client = Hexabase.instance;
-      await client.workspace.all();
-      expect(true, false);
-    } catch (e) {
-      expect(
-          e.toString(),
-          contains(
-              'You must initialize the Hexabase instance before calling Hexabase.instance'));
-    }
-  });
 
   test('Get all workspaces', () async {
     var client = Hexabase.instance;
-    var workspaces = await client.workspace.all();
+    var workspaces = await client.workspaces();
+    print(workspaces.map((e) => e.id));
     expect(workspaces[0].id, isNot(''));
     expect(workspaces[0].name, isNot(''));
   });