+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# CHANGELOG
+
+## 0.1.0
+
+- Create application
+- Update application name
+- Delete application
+
 ## 0.0.2
 
 Change initializing method.